Add really basic benchmark
This patch adds a really bad benchmark by refactoring the entire program
to a library module, except the final image output. But something is
better than nothing.

The benchmark rebuilds the scene every time, and the RNG's seed is not
fixed, so the results are noisy. But this can all be improved with time,
and at least now there is some method to monitor performance.