Use rand::thread_rng()
This patch reworks the code to use thread_rng instead of custom code.
This does remove some flexibility to use a pre-seeded RNG, and it's not
clear it is any faster given that threaded_rng() is called in every
function call that needs an RNG.

Passing RNG references to the functions that need it was not immediately
possible because it made the Material trait non-object safe. See
https://doc.rust-lang.org/error-index.html#method-has-generic-type-parameters

On WAVE:
```
     Running target\release\deps\benchmark-5f401f6c620b9427.exe
tracescene/10x10x4      time:   [586.13 us 588.06 us 590.27 us]
Found 12 outliers among 100 measurements (12.00%)
  2 (2.00%) low mild
  3 (3.00%) high mild
  7 (7.00%) high severe
```