Implement BVH again based on pbrt and indexing
This patch re-implements the BVH using a single Vec container for the
nodes and indices inside the nodes. This works around (with?) the borrow
checker, more or less.

This patch implements BVH traversal based on pbrt and rips out the
broken BVH construction code. A new implementation will come in a later
patch.